refactor(scripts): #72 2.2-7 — extract ECB/Fed/NBS/METI/StatBureau blocks
Move the final five source blocks to dedicated modules:

  scripts/validate/ecb.py           254 LOC
  scripts/validate/fed.py           282 LOC
  scripts/validate/nbs.py           199 LOC
  scripts/validate/meti.py           92 LOC
  scripts/validate/stat_bureau.py   156 LOC

The validator drops 2,024 → 1,152 LOC (−872). With this slice all
14 source-specific planner+runner blocks now live in
scripts/validate/<source>.py — the validator retains only its CLI,
dispatcher, report writer, and per-source `_run_<provider>` wrappers
that drive the live-execution flow (these are tightly coupled to
argparse + report rendering and stay until 2.2-9).

ECB, Fed, NBS keep their distinct `XxxProbe` dataclasses (each
carries source-specific call coordinates). METI and Statistics
Bureau JP both reuse the generic `Probe` shape — their probe
identity is just `name + path + params`, so a custom dataclass
would be redundant.

Test-compat note: 11 scaffold tests reach into ECB / Fed / NBS
private helpers via spec_from_file_location
(`validator._diff_ecb_observation`, `validator._try_parse_ecb`,
`validator._try_project_fomc_entry`,
`validator._try_project_release_entry`,
`validator._try_project_nbs_entry`). The validator's import block
re-exports those private names — same shim pattern slice 2.2-5
established for BLS/BEA/Census.

Tests: 9 scaffold tests at 135 passed; broader subset at 695
passed. Manual `--provider {ecb,fed,nbs,meti,stat-bureau-jp}`
dry-runs all plan correctly.

Refs #72.
