+++ scripts/validate_calendar_acquisition.py
@@ -130,60 +130,38 @@ from scripts.validate.nar import (  # noqa: E402
     run_nar_probe,
 )
 
-# Scaffold under test — we import and exercise exactly what production uses.
-from ingestion.calendar.ecb_api import (  # noqa: E402
-    INDICATOR_REGISTRY as ECB_INDICATOR_REGISTRY,
-    parse_observation as parse_ecb_observation,
+from scripts.validate.ecb import (  # noqa: E402
+    ECBClient,
+    ECBProbe,
+    ECB_INDICATOR_REGISTRY,
+    ECB_OBS_EXPECTED_FIELDS,
+    _diff_ecb_observation,
+    _try_parse_ecb,
+    plan_ecb_probes,
+    run_ecb_probe,
 )
-from ingestion.calendar.ecb_api.fetcher import (  # noqa: E402
-    _collapse_to_rate_changes as _ecb_collapse_to_rate_changes,
+from scripts.validate.fed import (  # noqa: E402
+    FED_INDICATOR_REGISTRY,
+    FedProbe,
+    _try_project_fomc_entry,
+    _try_project_release_entry,
+    plan_fed_probes,
+    run_fed_probe,
 )
-from ingestion.timeseries.sdmx._types import SDMXObservation  # noqa: E402
-from ingestion.timeseries.sdmx.providers.ecb import ECBClient  # noqa: E402
-
-from ingestion.calendar.fed_api import (  # noqa: E402
-    FED_CALENDAR_JSON_URL,
-    FOMC_CALENDAR_URL,
-    FomcMeetingEntry,
-    FedReleaseEntry,
-    INDICATOR_REGISTRY as FED_INDICATOR_REGISTRY,
-    fetch_fed_calendar_json,
-    fetch_fomc_calendar_html,
-    meeting_entry_to_records,
-    parse_fed_calendar_json,
-    parse_fomc_calendar_html,
-    release_entry_to_records,
+from scripts.validate.nbs import (  # noqa: E402
+    NBS_INDICATOR_REGISTRY,
+    NBSProbe,
+    _try_project_nbs_entry,
+    plan_nbs_probes,
+    run_nbs_probe,
 )
-
-from ingestion.calendar.nbs_api import (  # noqa: E402
-    INDICATOR_REGISTRY as NBS_INDICATOR_REGISTRY,
-    NBSReleaseEntry,
-    discover_nbs_calendar_url,
-    fetch_nbs_calendar_index_html,
-    fetch_nbs_yearly_calendar_html,
-    parse_nbs_calendar_html,
-    release_entry_to_records as nbs_release_entry_to_records,
+from scripts.validate.meti import (  # noqa: E402
+    plan_meti_probes,
+    run_meti_probe,
 )
-from ingestion.calendar.meti_api import (  # noqa: E402
-    ESTAT_RELEASE_CALENDAR_URL,
-    METI_RETAIL_PAGE_URL,
-    fetch_iip_release_calendar_html,
-    fetch_retail_schedule_html,
-    parse_iip_release_calendar_html,
-    parse_retail_schedule_html,
-    schedule_entry_to_records as meti_schedule_entry_to_records,
-)
-from ingestion.calendar.stat_bureau_api import (  # noqa: E402
-    ESTAT_STATS_DATA_URL,
-    INDICATOR_REGISTRY as STAT_BUREAU_INDICATOR_REGISTRY,
-    fetch_cpi_release_schedule_html,
-    fetch_estat_value_json,
-    fetch_lfs_release_schedule_html,
-    parse_cpi_release_schedule_html,
-    parse_estat_value_json,
-    parse_lfs_release_schedule_html,
-    schedule_entry_to_records as stat_bureau_schedule_entry_to_records,
-    time_code_for_month,
+from scripts.validate.stat_bureau import (  # noqa: E402
+    plan_stat_bureau_probes,
+    run_stat_bureau_probe,
 )
 
 
@@ -373,856 +351,6 @@ def _fmt_field_list(fields: list[str]) -> str:
         return "(none)"
     return ", ".join(f"`{f}`" for f in fields)
 
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# ECB probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-# SDMX observations arrive as a typed dataclass (:class:`SDMXObservation`)
-# rather than a raw upstream dict — the SDMX parser consumes the JSON
-# envelope at the boundary. "Field diff" therefore reports the parser-
-# facing attributes rather than upstream keys; upstream schema drift
-# surfaces as empty-results or type warnings, not UNKNOWN_OBSERVED.
-ECB_OBS_EXPECTED_FIELDS: frozenset[str] = frozenset({
-    "series_id", "date", "value", "dataflow",
-})
-
-
-@dataclass
-class ECBProbe:
-    """One ECB SDMX live-validation probe — a single series + date window.
-
-    Mirrors :class:`BLSProbe` / :class:`BEAProbe`. ECB's data path is
-    ``GET /service/data/{dataflow}/{key}`` — no auth, `jsondata` format,
-    ``startPeriod`` / ``endPeriod`` / ``lastNObservations`` query params.
-    """
-
-    name: str
-    series_id: str
-    dataflow_id: str
-    series_key: str
-    indicator: str
-    description: str
-    start_period: str     # ISO date
-    end_period: str       # ISO date
-
-
-def plan_ecb_probes() -> list[ECBProbe]:
-    """One probe per entry in the ECB calendar ``INDICATOR_REGISTRY``.
-
-    Each probe pulls roughly two years of business-daily observations
-    for the series — enough to cover multiple Governing Council
-    decisions in any realistic recent window. The runner reports both
-    the raw count (~500/year of business days) and the collapsed rate-
-    change count (~3–6 per window) so an operator can eyeball whether
-    the fetcher would project the expected signal from the firehose.
-
-    ECB requires no auth, so there's no ``api_key`` guard to bypass.
-    Three probes per run — under the endpoint's unspecified but
-    generous rate limit (we throttle via the SDMX client's retry
-    shape regardless).
-    """
-    today = datetime.now(timezone.utc).date()
-    # Calendar-day rollback rather than calendar-year so a leap year
-    # doesn't change the window size; 730 days fully covers two GC
-    # cycles (~8 scheduled decisions per year).
-    two_years_ago = today - timedelta(days=730)
-    probes: list[ECBProbe] = []
-    for series_id, spec in ECB_INDICATOR_REGISTRY.items():
-        probes.append(
-            ECBProbe(
-                name=f"{spec.indicator.lower()}_two_year_window",
-                series_id=series_id,
-                dataflow_id=spec.dataflow_id,
-                series_key=spec.series_key,
-                indicator=spec.indicator,
-                description=spec.title,
-                start_period=two_years_ago.isoformat(),
-                end_period=today.isoformat(),
-            )
-        )
-    return probes
-
-
-def _diff_ecb_observation(obs: SDMXObservation) -> RowDiff:
-    """Field diff for one SDMX observation.
-
-    "Observed" means the attributes present + non-null after the SDMX
-    parser has run — ECB's JSON envelope is already consumed at that
-    boundary. Missing-field / type warnings surface parser mis-
-    mappings rather than upstream schema drift; a fully empty
-    response from ECB (series key retired upstream) is caught at the
-    runner level, not here.
-    """
-    present = {
-        "series_id": obs.series_id,
-        "date":      obs.date,
-        "value":     obs.value,
-        "dataflow":  obs.dataflow,
-    }
-    observed = {k for k, v in present.items() if v not in (None, "")}
-    diff = RowDiff(
-        observed_fields=sorted(observed),
-        read_by_parser=sorted(observed & ECB_OBS_EXPECTED_FIELDS),
-        ignored_by_parser=[],   # no ambient upstream fields — dataclass is closed
-        unknown_observed=[],    # likewise
-        missing_expected=sorted(ECB_OBS_EXPECTED_FIELDS - observed),
-    )
-    if obs.value is None:
-        diff.type_warnings.append(
-            "value=None — rate-level series should always carry a numeric"
-        )
-    elif not isinstance(obs.value, (int, float)):
-        diff.type_warnings.append(
-            f"value is {type(obs.value).__name__}={obs.value!r} — "
-            f"SDMX parser expects numeric"
-        )
-    if not obs.date:
-        diff.type_warnings.append("date empty — parser would drop this row")
-    return diff
-
-
-def _try_parse_ecb(obs: SDMXObservation) -> tuple[bool, str]:
-    """Dry-run the calendar-side ECB projector on one observation."""
-    spec = ECB_INDICATOR_REGISTRY.get(obs.series_id)
-    if spec is None:
-        return False, f"no ECB calendar spec for series_id={obs.series_id!r}"
-    try:
-        raw, event = parse_ecb_observation(
-            obs, snapshot_epoch_ms=1_700_000_000_000, spec=spec,
-        )
-        return True, (
-            f"ok indicator={spec.indicator} event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def run_ecb_probe(client: ECBClient, probe: ECBProbe) -> ProbeResult:
-    """Execute one ECB probe and return a populated :class:`ProbeResult`.
-
-    Runs via ``ECBClient.get_data`` — reuses the production SDMX
-    transport so the probe exercises exactly the path a recurring
-    fetch would. An empty response lands as ``ok`` with ``row_count=0``
-    plus a note that the series key may have been retired upstream
-    (the signal we need for P6 parity not to silently drift).
-    """
-    generic = Probe(
-        name=probe.name,
-        path=(
-            f"GET https://data-api.ecb.europa.eu/service/data/"
-            f"{probe.dataflow_id}/{probe.series_key}"
-        ),
-        description=probe.description,
-        expected_shape="list[SDMXObservation] after client parse",
-        expected_fields=ECB_OBS_EXPECTED_FIELDS,
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = (
-        f"{generic.path}?format=jsondata"
-        f"&startPeriod={probe.start_period}"
-        f"&endPeriod={probe.end_period}"
-    )
-
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        observations = client.get_data(
-            probe.dataflow_id,
-            probe.series_key,
-            series_id=probe.series_id,
-            start_period=probe.start_period,
-            end_period=probe.end_period,
-            limit=0,  # no lastNObservations cap
-        )
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    result.status = "ok"
-    result.row_count = len(observations)
-    if not observations:
-        result.notes.append(
-            "ECB returned zero observations for this series + window "
-            "— series key may have been retired upstream"
-        )
-        return result
-
-    # FM-lane publishes the same level every business day until the
-    # Governing Council moves it. Expose both counts so the report
-    # makes the ~500-per-2-years business-daily noise vs the ~8-per-
-    # 2-years policy-move signal explicit.
-    changes = _ecb_collapse_to_rate_changes(observations, prior_value=None)
-    result.notes.append(
-        f"raw observations: {len(observations)} | rate changes after "
-        f"collapse: {len(changes)}"
-    )
-
-    # Newest-first via lexical date sort (ISO YYYY-MM-DD).
-    sorted_obs = sorted(observations, key=lambda o: o.date, reverse=True)
-    sample = sorted_obs[0]
-    result.sample_row = {
-        "series_id": sample.series_id,
-        "date":      sample.date,
-        "value":     sample.value,
-        "dataflow":  sample.dataflow,
-    }
-    result.field_diff = _diff_ecb_observation(sample)
-
-    # Unique rate-level tally — makes the policy-move cadence visible
-    # inside the window, and flags oddities (sub-zero prints, unusual
-    # precision) the parser should be ready for.
-    value_counter: Counter = Counter()
-    for obs in observations:
-        value_counter[repr(obs.value)] += 1
-    result.enum_counters = {"value": value_counter}
-
-    # Dry-parse the collapsed-changes subset rather than the business-
-    # daily firehose — the fetcher projects only these rows, so
-    # they're the set whose parse failure would matter.
-    target = changes if changes else sorted_obs
-    sample_n = min(10, len(target))
-    result.parse_attempts = sample_n
-    for obs in target[:sample_n]:
-        ok, msg = _try_parse_ecb(obs)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# Fed probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-# Fed publishes no authenticated calendar API. ``fomccalendars.htm``
-# is an HTML scrape; ``/json/calendar.json`` (Beige Book / H.4.1 / H.8
-# alongside FOMC meetings and speeches) is a JSON feed. The probe
-# runner doesn't do a per-field observation diff — the parser consumes
-# the payload at the boundary. Upstream drift surfaces as zero parsed
-# entries (parser raises) or non-empty ``row_issues`` (partial row-
-# level failures after a title match).
-
-
-@dataclass
-class FedProbe:
-    """One Fed live-validation probe — a single HTML surface.
-
-    Fed has two scrape surfaces, each driven by its own fetcher +
-    parser. ``source`` selects the dispatch branch inside
-    :func:`run_fed_probe`; ``url`` is informational (printed in the
-    dry-run plan + the report's request-path line).
-    """
-
-    name: str
-    source: str          # "fomc_calendar" | "releasedates"
-    url: str
-    description: str
-
-
-def plan_fed_probes() -> list[FedProbe]:
-    """Two probes — one per Fed calendar surface.
-
-    The FOMC HTML calendar carries ~48 meetings across a 6-year rolling
-    window (3 years past + current + 2 years forward) in a stable
-    panel-per-year layout; ``/json/calendar.json`` carries the full
-    rolling calendar — weekly H.4.1 / H.8 (one entry per month, days
-    comma-separated), ~8/year Beige Book, plus FOMC meetings, speeches,
-    and testimony we filter out. Each probe issues exactly one HTTP
-    request per run — no fan-out.
-    """
-    return [
-        FedProbe(
-            name="fomc_calendar",
-            source="fomc_calendar",
-            url=FOMC_CALENDAR_URL,
-            description=(
-                "FOMC meeting calendar — 6-year rolling panel of meeting "
-                "dates + SEP markers"
-            ),
-        ),
-        FedProbe(
-            name="fed_releasedates",
-            source="releasedates",
-            url=FED_CALENDAR_JSON_URL,
-            description=(
-                "Fed calendar JSON — rolling Beige Book / H.4.1 / H.8 "
-                "schedule"
-            ),
-        ),
-    ]
-
-
-def _try_project_fomc_entry(entry: FomcMeetingEntry) -> tuple[bool, str]:
-    """Dry-run the FOMC meeting → calendar-record projection."""
-    try:
-        raw, event = meeting_entry_to_records(
-            entry, snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok indicator=FOMC_RATE closing={entry.closing_date.isoformat()} "
-            f"event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _try_project_release_entry(entry: FedReleaseEntry) -> tuple[bool, str]:
-    """Dry-run the release-entry → calendar-record projection."""
-    try:
-        raw, event = release_entry_to_records(
-            entry, snapshot_epoch_ms=1_700_000_000_000,
-        )
-        return True, (
-            f"ok indicator={entry.series_id} date={entry.release_date} "
-            f"event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def _run_fomc_probe(result: ProbeResult, fomc_fetcher) -> ProbeResult:
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        html = fomc_fetcher()
-        entries = parse_fomc_calendar_html(html)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    result.row_count = len(entries)
-    if not entries:
-        # ``fetch_fed_calendar`` raises ``FomcCalendarParseError`` on
-        # this exact condition — a 200 interstitial or DOM drift that
-        # leaves ``parse_fomc_calendar_html`` with no matched rows.
-        # The probe must mirror that loud-fail semantics, otherwise a
-        # drifted upstream produces a green ``ok`` card and the report
-        # summary can claim the acquisition layer matches expectations
-        # while production would refuse to fetch.
-        result.status = "http_error"
-        result.notes.append(
-            "zero FOMC meetings parsed — DOM drift or access-denied "
-            "interstitial (production fetcher raises on this path)"
-        )
-        return result
-    result.status = "ok"
-
-    # Newest-first: closing_date is a date; sort desc.
-    ordered = sorted(entries, key=lambda e: e.closing_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "year":         sample.year,
-        "month_name":   sample.month_name,
-        "date_cell":    sample.date_cell,
-        "closing_date": sample.closing_date.isoformat(),
-        "has_sep":      sample.has_sep,
-    }
-
-    # Year-span + SEP tally — report the coverage surface so an
-    # operator eyeballing the card can see at a glance whether the
-    # panel still spans the expected ~6-year window and how many SEP
-    # (dot-plot) meetings sit inside it.
-    years = sorted({e.year for e in entries})
-    sep_count = sum(1 for e in entries if e.has_sep)
-    result.notes.append(
-        f"year span: {years[0]} → {years[-1]} "
-        f"({len(years)} distinct years) | SEP meetings: {sep_count}"
-    )
-
-    year_counter: Counter = Counter()
-    for e in entries:
-        year_counter[repr(e.year)] += 1
-    result.enum_counters = {"year": year_counter}
-
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_fomc_entry(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def _run_releasedates_probe(result: ProbeResult, releasedates_fetcher) -> ProbeResult:
-    import time as _time
-    t0 = _time.monotonic()
-    row_issues: list[str] = []
-    try:
-        text = releasedates_fetcher()
-        entries = parse_fed_calendar_json(text, row_issues=row_issues)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    # ``parse_fed_calendar_json`` already raises on both empty-match
-    # and all-matches-failed paths (see the module for the exact
-    # guards), so ``entries`` is guaranteed non-empty once we reach
-    # here — the general exception branch above carries the loud-
-    # fail signal into ``status=http_error`` for both failure shapes.
-    result.status = "ok"
-    result.row_count = len(entries)
-    if row_issues:
-        # Surface partial-row failures exactly as the fetcher's summary
-        # does — the probe report is the debugging surface for whoever
-        # investigates a drift signal, so it shouldn't hide this.
-        for issue in row_issues[:5]:
-            result.notes.append(f"row_issue: {issue}")
-        if len(row_issues) > 5:
-            result.notes.append(
-                f"... {len(row_issues) - 5} more row_issues suppressed"
-            )
-
-    ordered = sorted(entries, key=lambda e: e.release_date, reverse=True)
-    sample = ordered[0]
-    result.sample_row = {
-        "series_id":          sample.series_id,
-        "release_title":      sample.release_title,
-        "release_date":       sample.release_date,
-        "release_time_local": sample.release_time_local,
-        "event_time_utc":     sample.event_time_utc,
-    }
-
-    by_indicator: Counter = Counter()
-    for e in entries:
-        by_indicator[e.series_id] += 1
-    result.enum_counters = {"series_id": by_indicator}
-    result.notes.append(
-        "entries by indicator: "
-        + ", ".join(f"{k}={v}" for k, v in by_indicator.most_common())
-    )
-
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_release_entry(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-def run_fed_probe(
-    probe: FedProbe,
-    *,
-    fomc_fetcher=fetch_fomc_calendar_html,
-    releasedates_fetcher=fetch_fed_calendar_json,
-) -> ProbeResult:
-    """Execute one Fed probe and return a populated :class:`ProbeResult`.
-
-    Unlike the API-based probes, Fed consumes public web responses
-    locally — HTML for FOMC calendar, JSON for the release feed. The
-    runner drives the production fetcher + parser; ``fomc_fetcher``
-    and ``releasedates_fetcher`` are seams tests inject to feed fixture
-    payloads without hitting ``federalreserve.gov``. ``auth_missing``
-    is not reachable — both surfaces are public.
-    """
-    generic = Probe(
-        name=probe.name,
-        path=f"GET {probe.url}",
-        description=probe.description,
-        expected_shape="list[FomcMeetingEntry] / list[FedReleaseEntry]",
-        expected_fields=frozenset(),  # HTML/JSON surface — diff not meaningful
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-    result.request_path = generic.path
-
-    if probe.source == "fomc_calendar":
-        return _run_fomc_probe(result, fomc_fetcher)
-    if probe.source == "releasedates":
-        return _run_releasedates_probe(result, releasedates_fetcher)
-    raise ValueError(f"unknown Fed probe source: {probe.source!r}")
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# NBS probe plan + runner
-# ──────────────────────────────────────────────────────────────────────────
-
-# Single-page HTML scrape per run — the NBS yearly calendar is one
-# article listing every registered indicator's 12-month (or
-# quarterly) release schedule. Probe coverage is therefore breadth
-# (how many indicators matched) rather than depth (per-indicator
-# surfaces). NBS is documented as the highest-risk upstream on this
-# issue: HTTP-only, HTML-fragile, frequent timeouts from non-CN IPs.
-# The probe runner catches any exception cleanly — a single-probe
-# timeout surfaces as a loud ``http_error`` card rather than crashing
-# the whole run; on the current single-indicator registry that's
-# also the only probe in flight, but the tolerance shape is in place
-# for when P5c-live expands the whitelist.
-
-
-@dataclass
-class NBSProbe:
-    """One NBS yearly-calendar live-validation probe."""
-
-    name: str
-    year: int
-    description: str
-
-
-def plan_nbs_probes() -> list[NBSProbe]:
-    """One probe per run — the yearly calendar article for the current
-    UTC year. The article lists every registered indicator together,
-    so a single fetch validates the breadth of
-    :data:`INDICATOR_REGISTRY` coverage.
-    """
-    year = datetime.now(timezone.utc).year
-    return [
-        NBSProbe(
-            name=f"nbs_yearly_calendar_{year}",
-            year=year,
-            description=(
-                f"NBS yearly release calendar article for {year} — covers "
-                f"every registered indicator in one fetch"
-            ),
-        )
-    ]
-
-
-def _try_project_nbs_entry(entry: NBSReleaseEntry) -> tuple[bool, str]:
-    """Dry-run the NBS calendar projection on one entry."""
-    try:
-        raw, event = nbs_release_entry_to_records(
-            entry,
-            snapshot_epoch_ms=1_700_000_000_000,
-            calendar_url="https://example.test/nbs_fixture",
-        )
-        return True, (
-            f"ok indicator={entry.indicator} "
-            f"date={entry.year}-{entry.month:02d}-{entry.day:02d} "
-            f"event_id={raw.provider_event_id[:10]}…"
-        )
-    except Exception as exc:
-        return False, f"{type(exc).__name__}: {exc}"
-
-
-def run_nbs_probe(
-    probe: NBSProbe,
-    *,
-    index_fetcher=fetch_nbs_calendar_index_html,
-    html_fetcher=fetch_nbs_yearly_calendar_html,
-) -> ProbeResult:
-    """Execute one NBS probe and return a populated :class:`ProbeResult`.
-
-    Auto-discovers the calendar URL via the index page, fetches the
-    article, parses entries, and dry-projects each through
-    ``release_entry_to_records``. ``index_fetcher`` + ``html_fetcher``
-    seams let tests feed fixture HTML without touching
-    ``stats.gov.cn``.
-
-    NBS upstream is notorious for intermittent timeouts from outside
-    CN; the single ``except Exception`` branch carries every failure
-    mode (index timeout, article timeout, parse raise) into a
-    ``http_error`` card rather than letting the exception propagate
-    and kill a multi-probe run (forward-compatible for P5c-live's
-    whitelist expansion).
-    """
-    generic = Probe(
-        name=probe.name,
-        path=(
-            f"GET http://www.stats.gov.cn/english/PressRelease/"
-            f"ReleaseCalendar/ → {probe.year} article"
-        ),
-        description=probe.description,
-        expected_shape="list[NBSReleaseEntry] after HTML parse",
-        expected_fields=frozenset(),  # HTML surface — diff not meaningful
-    )
-    result = ProbeResult(probe=generic, status="skipped")
-
-    import time as _time
-    t0 = _time.monotonic()
-    try:
-        calendar_url = discover_nbs_calendar_url(
-            probe.year, index_fetcher=index_fetcher,
-        )
-        result.request_path = f"GET {calendar_url}"
-        html = html_fetcher(calendar_url)
-        entries = parse_nbs_calendar_html(html, year_override=probe.year)
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(f"{type(exc).__name__}: {exc}")
-        result.notes.append(
-            "NBS upstream is the highest-risk source on this issue "
-            "(HTTP-only, HTML-fragile, frequent non-CN timeouts). "
-            "Transient failures here are expected — retry the probe "
-            "before treating it as a genuine drift signal."
-        )
-        return result
-    result.http_elapsed_ms = (_time.monotonic() - t0) * 1000
-
-    result.row_count = len(entries)
-    if not entries:
-        # The production fetcher raises ``NBSCalendarParseError`` on
-        # this exact path; mirror the loud-fail shape per the Codex
-        # review lesson from P4b — a green card on zero entries would
-        # let an interstitial slip through.
-        result.status = "http_error"
-        result.notes.append(
-            "zero NBS entries parsed — DOM drift or interstitial "
-            "(production fetcher raises on this path)"
-        )
-        return result
-    result.status = "ok"
-
-    # Newest-first by (year, month, day).
-    ordered = sorted(
-        entries,
-        key=lambda e: (e.year, e.month, e.day),
-        reverse=True,
-    )
-    sample = ordered[0]
-    result.sample_row = {
-        "year":               sample.year,
-        "month":              sample.month,
-        "day":                sample.day,
-        "indicator":          sample.indicator,
-        "release_time_local": sample.release_time_local,
-        "date_cell":          sample.date_cell,
-    }
-
-    # Per-indicator tally. The yearly calendar is the one place every
-    # registered indicator converges, so the count-per-indicator is
-    # the best drift signal: if PPI / GDP / etc. (P5c-live additions)
-    # land with zero matches, their label_fragment drifted off the
-    # page's actual headers.
-    by_indicator: Counter = Counter()
-    for e in entries:
-        by_indicator[e.indicator] += 1
-    result.enum_counters = {"indicator": by_indicator}
-    result.notes.append(
-        "entries by indicator: "
-        + ", ".join(f"{k}={v}" for k, v in by_indicator.most_common())
-    )
-
-    # Dry-project up to 10 newest entries — the set whose projection
-    # failure would matter most for an imminent release.
-    sample_n = min(10, len(ordered))
-    result.parse_attempts = sample_n
-    for entry in ordered[:sample_n]:
-        ok, msg = _try_project_nbs_entry(entry)
-        if ok:
-            result.parse_successes += 1
-        else:
-            if len(result.parse_error_samples) < 3:
-                result.parse_error_samples.append(msg)
-    return result
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# METI probes
-# ──────────────────────────────────────────────────────────────────────────
-
-
-def plan_meti_probes() -> list[Probe]:
-    return [
-        Probe(
-            name="meti_iip_release_calendar",
-            path=ESTAT_RELEASE_CALENDAR_URL,
-            description="e-Stat IIP release calendar (toukei_cd 00550300)",
-            expected_shape="HTML rows -> MetiScheduleEntry",
-            params={"surface": "iip"},
-        ),
-        Probe(
-            name="meti_retail_schedule",
-            path=METI_RETAIL_PAGE_URL,
-            description="METI Current Survey of Commerce next-release sentence",
-            expected_shape="HTML page -> MetiScheduleEntry",
-            params={"surface": "retail"},
-        ),
-    ]
-
-
-def run_meti_probe(probe: Probe) -> ProbeResult:
-    result = ProbeResult(probe=probe, status="ok", request_path=probe.path)
-    started = datetime.now(timezone.utc)
-    try:
-        if probe.name == "meti_iip_release_calendar":
-            today = started.date()
-            html_text = fetch_iip_release_calendar_html(
-                today - timedelta(days=14),
-                today + timedelta(days=90),
-            )
-            entries = parse_iip_release_calendar_html(html_text)
-        elif probe.name == "meti_retail_schedule":
-            html = fetch_retail_schedule_html()
-            entries = [parse_retail_schedule_html(html)]
-        else:
-            raise ValueError(f"unknown METI probe: {probe.name}")
-
-        result.row_count = len(entries)
-        if entries:
-            first = entries[0]
-            result.sample_row = {
-                "indicator": first.indicator,
-                "reference_date": first.reference_date.isoformat(),
-                "release_date": first.release_date.isoformat(),
-                "release_time_local": first.release_time_local,
-            }
-        result.enum_counters = {
-            "indicator": Counter(e.indicator for e in entries),
-        }
-        sample_n = min(10, len(entries))
-        result.parse_attempts = sample_n
-        for entry in entries[:sample_n]:
-            try:
-                meti_schedule_entry_to_records(entry, snapshot_epoch_ms=1)
-                result.parse_successes += 1
-            except Exception as exc:
-                if len(result.parse_error_samples) < 3:
-                    result.parse_error_samples.append(str(exc))
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(str(exc))
-    finally:
-        result.http_elapsed_ms = (
-            datetime.now(timezone.utc) - started
-        ).total_seconds() * 1000
-    return result
-
-
-# ──────────────────────────────────────────────────────────────────────────
-# Statistics Bureau / e-Stat probes
-# ──────────────────────────────────────────────────────────────────────────
-
-
-def plan_stat_bureau_probes() -> list[Probe]:
-    cpi_ref = date(2026, 3, 1)
-    lfs_ref = date(2026, 2, 1)
-    return [
-        Probe(
-            name="stat_bureau_cpi_schedule",
-            path="https://www.stat.go.jp/english/data/cpi/1582.htm",
-            description="CPI release schedule — Japan column",
-            expected_shape="HTML table rows -> StatBureauScheduleEntry",
-            params={"surface": "cpi"},
-        ),
-        Probe(
-            name="stat_bureau_lfs_schedule",
-            path="https://www.stat.go.jp/english/data/roudou/1543.htm",
-            description="Labour Force Survey release schedule — Basic tabulation",
-            expected_shape="HTML table rows -> StatBureauScheduleEntry",
-            params={"surface": "lfs"},
-        ),
-        Probe(
-            name="stat_bureau_core_cpi_value",
-            path=ESTAT_STATS_DATA_URL,
-            description="e-Stat Core CPI YoY scalar value",
-            expected_shape="GET_STATS_DATA.DATA_INF.VALUE scalar",
-            params={
-                "statsDataId": STAT_BUREAU_INDICATOR_REGISTRY["CORE_CPI"].stats_data_id,
-                "indicator": "CORE_CPI",
-                "reference": cpi_ref.isoformat(),
-                "cdTime": time_code_for_month(cpi_ref),
-            },
-        ),
-        Probe(
-            name="stat_bureau_unemployment_value",
-            path=ESTAT_STATS_DATA_URL,
-            description="e-Stat Unemployment Rate scalar value",
-            expected_shape="GET_STATS_DATA.DATA_INF.VALUE scalar",
-            params={
-                "statsDataId": STAT_BUREAU_INDICATOR_REGISTRY["UNEMPLOYMENT_RATE"].stats_data_id,
-                "indicator": "UNEMPLOYMENT_RATE",
-                "reference": lfs_ref.isoformat(),
-                "cdTime": time_code_for_month(lfs_ref),
-            },
-        ),
-    ]
-
-
-def run_stat_bureau_probe(probe: Probe) -> ProbeResult:
-    result = ProbeResult(probe=probe, status="ok", request_path=probe.path)
-    started = datetime.now(timezone.utc)
-    try:
-        if probe.name == "stat_bureau_cpi_schedule":
-            html = fetch_cpi_release_schedule_html()
-            entries = parse_cpi_release_schedule_html(html)
-        elif probe.name == "stat_bureau_lfs_schedule":
-            html = fetch_lfs_release_schedule_html()
-            entries = parse_lfs_release_schedule_html(html)
-        elif probe.name in {
-            "stat_bureau_core_cpi_value",
-            "stat_bureau_unemployment_value",
-        }:
-            app_id = os.getenv("ESTAT_APP_ID", "").strip()
-            if not app_id:
-                result.status = "auth_missing"
-                result.notes.append("ESTAT_APP_ID not set")
-                return result
-            indicator = str(probe.params["indicator"])
-            reference = date.fromisoformat(str(probe.params["reference"]))
-            spec = STAT_BUREAU_INDICATOR_REGISTRY[indicator]
-            data = fetch_estat_value_json(spec, reference, app_id=app_id)
-            value = parse_estat_value_json(
-                data,
-                indicator=indicator,
-                reference=reference,
-            )
-            result.row_count = 1
-            result.sample_row = {
-                "indicator": indicator,
-                "reference_date": value.reference_date.isoformat(),
-                "time_code": value.time_code,
-                "actual": value.actual,
-                "unit": value.unit,
-                "attrs": value.attrs,
-            }
-            result.parse_attempts = 1
-            result.parse_successes = 1
-            result.enum_counters = {
-                "indicator": Counter({indicator: 1}),
-                "unit": Counter({value.unit: 1}),
-            }
-            return result
-        else:
-            raise ValueError(f"unknown Statistics Bureau probe: {probe.name}")
-
-        result.row_count = len(entries)
-        if entries:
-            first = entries[0]
-            result.sample_row = {
-                "indicator": first.indicator,
-                "reference_date": first.reference_date.isoformat(),
-                "release_date": first.release_date.isoformat(),
-                "release_time_local": first.release_time_local,
-            }
-        result.enum_counters = {
-            "indicator": Counter(e.indicator for e in entries),
-        }
-        sample_n = min(10, len(entries))
-        result.parse_attempts = sample_n
-        for entry in entries[:sample_n]:
-            try:
-                stat_bureau_schedule_entry_to_records(
-                    entry,
-                    snapshot_epoch_ms=1,
-                )
-                result.parse_successes += 1
-            except Exception as exc:
-                if len(result.parse_error_samples) < 3:
-                    result.parse_error_samples.append(str(exc))
-    except Exception as exc:
-        result.status = "http_error"
-        result.notes.append(str(exc))
-    finally:
-        result.http_elapsed_ms = (
-            datetime.now(timezone.utc) - started
-        ).total_seconds() * 1000
-    return result
 
 
 # ──────────────────────────────────────────────────────────────────────────
